sales order cancel api created

+++ app/Modules/Sales/Http/Controllers/SalesController.php
@@ -14,10 +14,13 @@ use App\Modules\Sales\Http\Requests\CreateRequest;
 use App\Modules\Sales\Http\Requests\MailRequest;
 use App\Modules\Sales\Http\Requests\StatusRequest;
 use App\Modules\Sales\Http\Requests\UpdateRequest;
+use App\Modules\Sales\Http\Resources\RecipePartialOrder;
 use App\Modules\Sales\Http\Resources\SalesOrder as SalesOrderResource;
+use App\Modules\Sales\Models\Delivery;
 use App\Modules\Sales\Models\SalesOrder;
 use App\Modules\Sales\Models\SalesOrderRecipe;
 use App\Modules\Sales\Repositories\CompanyRepository;
+use App\Modules\Sales\Repositories\DeliveryRepository;
 use App\Modules\Sales\Repositories\RecipePartialRepository;
 use App\Modules\Sales\Repositories\SalesOrderRepository;
 use App\Modules\Sales\Repositories\SalesRecipeRepository;
@@ -63,6 +66,11 @@ class SalesController extends Controller
 
     protected $threadColorRepository;
 
+    protected $stockRepository;
+
+    protected $deliveryRepository;
+
+
     /**
      * SalesController constructor.
      * @param SalesOrderRepository    $salesOrderRepository
@@ -70,19 +78,25 @@ class SalesController extends Controller
      * @param DesignDetailRepository  $designDetailRepository
      * @param RecipePartialRepository $recipePartialOrderRepository
      * @param ThreadColorRepository   $threadColorRepository
+     * @param StockRepository         $stockRepository
+     * @param DeliveryRepository      $deliveryRepository
      */
     public function __construct(
         SalesOrderRepository $salesOrderRepository,
         MasterRepository $masterRepository,
         DesignDetailRepository $designDetailRepository,
         RecipePartialRepository $recipePartialOrderRepository,
-        ThreadColorRepository $threadColorRepository
+        ThreadColorRepository $threadColorRepository,
+        StockRepository $stockRepository,
+        DeliveryRepository $deliveryRepository
     ) {
         $this->salesOrderRepository = $salesOrderRepository;
         $this->masterRepository = $masterRepository;
         $this->designDetailRepository = $designDetailRepository;
         $this->recipePartialOrderRepo = $recipePartialOrderRepository;
         $this->threadColorRepository = $threadColorRepository;
+        $this->stockRepository = $stockRepository;
+        $this->deliveryRepository = $deliveryRepository;
     }
 
 
@@ -130,7 +144,7 @@ class SalesController extends Controller
     private function createOrUpdateSalesDetails(SalesOrder $salesOrder, $input, $update) {
         $designDetail = $this->designDetailRepository->findBy('design_id',
             $input['design_id'], ['panno', 'additional_panno', 'reed']);
-        $salesOrder->load('designBeam.threadColor.thread');
+        $salesOrder->load(['designBeam.threadColor.thread', 'status']);
         $this->storeSalesOrderRecipes($salesOrder, $input, $designDetail, $update);
     }
 
@@ -146,21 +160,22 @@ class SalesController extends Controller
         $designDetail,
         $update
     ) {
-        if ($update) {
-            // remove old stock results
-            $salesOrder->orderStocks()->delete();
-        }
+        if ($salesOrder->status->code != MasterConstant::SO_MANUFACTURING) {
+            if ($update) {
+                // remove old stock results
+                $salesOrder->orderStocks()->delete();
+            }
 
-        foreach ($input['order_recipes'] as $items) {
-            $orderRecipeId = isset($items['id']) ? $items['id'] : null;
-            $orderRecipe = $salesOrder->orderRecipes()
-                                      ->updateOrCreate(['id' => $orderRecipeId], $items);
-            $items['status_id'] = $salesOrder->status_id;
-            $this->storeRecipeOrderQuantities($salesOrder, $orderRecipe, $items, $designDetail);
+            foreach ($input['order_recipes'] as $items) {
+                $orderRecipeId = isset($items['id']) ? $items['id'] : null;
+                $orderRecipe = $salesOrder->orderRecipes()
+                                          ->updateOrCreate(['id' => $orderRecipeId], $items);
+                $this->storeRecipeOrderQuantities($salesOrder, $orderRecipe, $items, $designDetail);
+            }
         }
 
         if ($update && isset($input['removed_order_recipes_id']) && !empty($input['removed_order_recipes_id'])) {
-            $this->destroyOrderRecipes($input['removed_order_recipes_id']);
+            $this->destroyOrderRecipes($input['removed_order_recipes_id'], $salesOrder);
         }
     }
 
@@ -185,7 +200,7 @@ class SalesController extends Controller
                 'order_recipe_id' => $orderRecipe->id,
                 'product_id'      => $quantityDetails['thread_color_id'],
                 'product_type'    => 'thread_color',
-                'status_id'       => $items['status_id'],
+                'status_id'       => $salesOrder->status_id,
                 'kg_qty'          => '-' . $formula->getTotalKgQty(ThreadType::WEFT,
                         $quantityDetails, $designDetail, $items['total_meters']),
             ];
@@ -196,7 +211,7 @@ class SalesController extends Controller
             'order_recipe_id' => $orderRecipe->id,
             'product_id'      => $salesOrder->designBeam->thread_color_id,
             'product_type'    => 'thread_color',
-            'status_id'       => $items['status_id'],
+            'status_id'       => $salesOrder->status_id,
             'kg_qty'          => '-' . $formula->getTotalKgQty(ThreadType::WARP,
                     $threadDetail, $designDetail, $items['total_meters']),
         ]);
@@ -206,12 +221,19 @@ class SalesController extends Controller
 
     /**
      * @param $orderRecipeIds
+     * @param $salesOrder
      */
-    private function destroyOrderRecipes($orderRecipeIds) {
+    private function destroyOrderRecipes($orderRecipeIds, $salesOrder) {
+
+        if ($salesOrder->status->code == MasterConstant::SO_MANUFACTURING) {
+            $partialOrderIds = $this->recipePartialOrderRepo->findIdByRecipeIds($orderRecipeIds);
+            $this->deliveryRepository->removeSinglePartialOrders($partialOrderIds, $salesOrder->id);
+            $this->recipePartialOrderRepo->removeByField('id', $partialOrderIds);
+        }
         // remove sales partial order stocks
-        (new StockRepository(new Container()))->removeByPartialOrderId($orderRecipeIds);
+        $this->stockRepository->removeByField('order_recipe_id', $orderRecipeIds);
         // remove sales partial orders
-        (new SalesRecipeRepository(new Container()))->removeById($orderRecipeIds);
+        (new SalesRecipeRepository(new Container()))->removeByField('id', $orderRecipeIds);
 
     }
 
@@ -261,8 +283,7 @@ class SalesController extends Controller
         try {
             DB::beginTransaction();
             $salesOrder->update($input);
-            $salesOrder->fresh();
-            $this->createOrUpdateSalesDetails($salesOrder, $input, true);
+            $this->createOrUpdateSalesDetails($salesOrder->refresh(), $input, true);
             DB::commit();
 
             return $this->sendResponse($this->makeResource($salesOrder),
@@ -345,6 +366,9 @@ class SalesController extends Controller
         }
     }
 
+    /**
+     * @return mixed
+     */
     private function totalMeterStatuses() {
         return $this->masterRepository->findWhereIn('code',
             [Master::SO_DELIVERED, Master::SO_MANUFACTURING], ['id', 'code'])
@@ -558,7 +582,13 @@ class SalesController extends Controller
 
         $status = $this->masterRepository->findByCode(MasterConstant::SO_CANCELED);
 
-        $salesOrder->orderStocks()->update(['status_id' => $status->id]);
+        $salesOrder->deliveries->each(function (Delivery $delivery) {
+            $delivery->partialOrders()->delete();
+        });
+
+        $salesOrder->deliveries()->delete();
+
+        $salesOrder->orderStocks()->delete();
 
         return $this->updateStatus($salesOrder, $status);
